plane-claude-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
aiohttp==3.10.0
pyyaml==6.0.1
python-dotenv==1.0.1
websockets==13.0
pydantic==2.9.0
pydantic-settings==2.6.0
tenacity==8.2.3
prometheus-client==0.20.0
